test: merge existing path tests
