+++ test/Makefile.autosetup
@@ -2,7 +2,6 @@ TEST_OBJS   = test/main.o \
 	      test/base64.o \
 	      test/buffer.o \
 	      test/md5.o \
-	      test/path.o \
 	      test/rfc2047.o \
 	      test/address.o \
 	      test/url.o \
