+++ test/main.c
@@ -36,12 +36,8 @@
   NEOMUTT_TEST_ITEM(test_md5_ctx)                                              \
   NEOMUTT_TEST_ITEM(test_md5_ctx_bytes)                                        \
   NEOMUTT_TEST_ITEM(test_addr_mbox_to_udomain)                                 \
-  NEOMUTT_TEST_ITEM(test_mutt_path_tidy_slash)                                 \
-  NEOMUTT_TEST_ITEM(test_mutt_path_tidy_dotdot)                                \
-  NEOMUTT_TEST_ITEM(test_mutt_path_tidy)                                       \
   NEOMUTT_TEST_ITEM(test_url)                                                  \
   NEOMUTT_TEST_ITEM(test_mutt_buffer_concat_path)                              \
-  NEOMUTT_TEST_ITEM(test_mutt_path_getcwd)
 
 /******************************************************************************
  * You probably don't need to touch what follows.
