+++ test/path/mutt_path_tidy_slash.c
@@ -29,7 +29,50 @@ void test_mutt_path_tidy_slash(void)
 {
   // bool mutt_path_tidy_slash(char *buf);
 
+  static const char *tests[][2] =
+  {
+    { NULL,                     NULL,            },
+    { "/",                      "/",             },
+    { "//",                     "/",             },
+    { "///",                    "/",             },
+    { "/apple/",                "/apple",        },
+    { "/apple//",               "/apple",        },
+    { "/apple///",              "/apple",        },
+    { "/apple/banana",          "/apple/banana", },
+    { "/apple//banana",         "/apple/banana", },
+    { "/apple///banana",        "/apple/banana", },
+    { "/apple/banana/",         "/apple/banana", },
+    { "/apple/banana//",        "/apple/banana", },
+    { "/apple/banana///",       "/apple/banana", },
+    { "//.///././apple/banana", "/apple/banana", },
+    { "/apple/.///././banana",  "/apple/banana", },
+    { "/apple/banana/.///././", "/apple/banana", },
+    { "/apple/banana/",         "/apple/banana", },
+    { "/apple/banana/.",        "/apple/banana", },
+    { "/apple/banana/./",       "/apple/banana", },
+    { "/apple/banana//",        "/apple/banana", },
+    { "/apple/banana//.",       "/apple/banana", },
+    { "/apple/banana//./",      "/apple/banana", },
+    { "////apple/banana",       "/apple/banana", },
+    { "/.//apple/banana",       "/apple/banana", },
+  };
+
   {
     TEST_CHECK(!mutt_path_tidy_slash(NULL));
+  }
+
+  {
+    char buf[64];
+    for (size_t i = 0; i < mutt_array_size(tests); i++)
+    {
+      mutt_str_strfcpy(buf, tests[i][0], sizeof(buf));
+      mutt_path_tidy_slash(buf);
+      if (!TEST_CHECK(mutt_str_strcmp(buf, tests[i][1]) == 0))
+      {
+        TEST_MSG("Input:    %s", tests[i][0]);
+        TEST_MSG("Expected: %s", tests[i][1]);
+        TEST_MSG("Actual:   %s", buf);
+      }
+    }
   }
 }
